"mission8 #5 Feat: 미션추가 api구현"

+++ src/main/java/umc/study/web/dto/ReviewRequestDTO.java
@@ -10,8 +10,8 @@ public class ReviewRequestDTO {
         @NotNull
         String title;
         @NotNull
-        Integer member_id;
+        Long member_id;
         @NotNull
-        Integer store_id;
+        Long store_id;
     }
 }
